requirements.txt
astroid==2.12.12
attrs==22.1.0
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
cycler==0.11.0
data-science-types==0.2.23
dill==0.3.5.1
idna==3.4
iniconfig==1.1.1
isort==5.10.1
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
matplotlib==3.1.1
mccabe==0.7.0
mypy==0.982
mypy-extensions==0.4.3
mysql-connector-python==8.0.17
numpy==1.17.2
packaging==21.3
pkg_resources==0.0.0
platformdirs==2.5.2
pluggy==1.0.0
protobuf==4.21.8
psycopg2==2.8.3
py==1.11.0
py-package-template==0.2.0
PyAudio==0.2.11
pycodestyle==2.9.1
pydub==0.23.1
pylint==2.15.4
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
scipy==1.3.1
six==1.16.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
typing_extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
